renaming config to constants, exporting getPicture()
