+++ packages/integrations/image/src/constants.ts
